requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
openai==1.51.0
httpx>=0.24.0,<0.28.0
pymupdf==1.24.10
pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
